\name{matfact-package}
\alias{matfact-package}
\docType{package}
\title{
matfact: Matrix Factorization Modeling
}
\description{
Implements a family of matrix factorization models via non-convex optimization. The models considered include low rank matrix linear regression, low rank matrix logistic regression, and robust principle component analysis.
}
\details{
\tabular{ll}{
Package: \tab matfact\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2016-08-05\cr
License: \tab GPL-2\cr
}
}

\author{
Xingguo Li, Enayat Ullah and Tuo Zhao \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}

\seealso{
\code{\link{matfact}}.
}

